Move CA generation onto CertificateAuthority
Breaks dependency on sun.security in Jetty94HttpServer
